requirements.txt
pyTelegramBotAPI==4.6
python-dotenv==0.20.0
mysql-connector-python==8.0.29
requests=2.28.1
loguru=0.6.0
pytz=2022.1